requirements.txt
snowflake-connector-python==4.1.0
snowflake-snowpark-python==1.42.0
python-dotenv==1.2.1
google-genai==1.52.0
google-adk==1.19.0